requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
audioread==3.0.1
census==0.8.24
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
datasets==3.5.0
decorator==5.2.1
dill==0.3.8
filelock==3.18.0
frozenlist==1.6.0
fsspec==2024.12.0
huggingface-hub==0.30.2
idna==3.10
jellyfish==1.2.0
joblib==1.4.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
numba==0.61.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
platformdirs==4.3.7
pooch==1.8.2
propcache==0.3.1
pyarrow==19.0.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
us==3.2.0
xxhash==3.5.0
yarl==1.20.0
